requirements.txt
arabic-reshaper==2.1.1
asgiref==3.2.10
cairocffi==1.2.0
CairoSVG==2.5.1
cffi==1.14.4
cssselect2==0.4.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.2
django-crispy-forms==1.9.2
django-import-export==2.4.0
et-xmlfile==1.0.1
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==2.0.1
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
pycparser==2.20
PyMySQL==0.10.1
PyPDF2==1.26.0
Pyphen==0.10.0
python-bidi==0.4.2
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
tinycss2==1.1.0
WeasyPrint==52.2
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
